Update authorized_key.py
##### SUMMARY
Clarification of one example:
I just wanted to mark a possible confusion in the example. If one would try the example with the "with_file" and "exclusive" option it would work. Then one could try to add more keys to this exact example and play around with it ans whoooosh... only one key will remain .. unexpectedly


##### ISSUE TYPE
- Docs Pull Request

+label: docsite_pr